Rewrote all positions and velocities to float, not working yet

+++ code/base_code/program/dsmc/cell.cpp
@@ -53,18 +53,18 @@ unsigned long Cell::prepare() {
     return collision_pairs;
 }
 
-void Cell::collide_molecules(double *v0, double *v1, const double &v_rel, Random *rnd) {
-    double vcmx  = 0.5*(v0[0] + v1[0]);
-    double vcmy  = 0.5*(v0[1] + v1[1]);
-    double vcmz  = 0.5*(v0[2] + v1[2]);
+void Cell::collide_molecules(float *v0, float *v1, const float &v_rel, Random *rnd) {
+    float vcmx  = 0.5*(v0[0] + v1[0]);
+    float vcmy  = 0.5*(v0[1] + v1[1]);
+    float vcmz  = 0.5*(v0[2] + v1[2]);
 
-    double cos_th = 1.0 - 2.0*rnd->next_double();      // Cosine and sine of
-    double sin_th = sqrt(1.0 - cos_th*cos_th);        // collision angle theta
-    double phi = 2*M_PI*rnd->next_double();
+    float cos_th = 1.0 - 2.0*rnd->next_double();      // Cosine and sine of
+    float sin_th = sqrt(1.0 - cos_th*cos_th);        // collision angle theta
+    float phi = 2*M_PI*rnd->next_double();
 
-    double vrelx = v_rel*cos_th;                   // Compute post-collision relative velocity
-    double vrely = v_rel*sin_th*cos(phi);
-    double vrelz = v_rel*sin_th*sin(phi);
+    float vrelx = v_rel*cos_th;                   // Compute post-collision relative velocity
+    float vrely = v_rel*sin_th*cos(phi);
+    float vrelz = v_rel*sin_th*sin(phi);
 
     v0[0] = vcmx + 0.5*vrelx;
     v0[1] = vcmy + 0.5*vrely;
@@ -81,7 +81,7 @@ int Cell::collide(Random *rnd) {
 
     if( num_molecules < 2 ) return 0;  // Skip to the next cell
 
-    double crm = vr_max;     // Current maximum relative speed
+    float crm = vr_max;     // Current maximum relative speed
 
 	//* Loop over total number of candidate collision pairs
     int collisions = 0;
@@ -94,9 +94,9 @@ int Cell::collide(Random *rnd) {
         ip1 = molecules[ip1];
 
 		//* Calculate pair's relative speed
-        double *v0 = &system->v[3*ip0];
-        double *v1 = &system->v[3*ip1];
-        double v_rel = sqrt(pow(v0[0] - v1[0],2) + pow(v0[1] - v1[1],2) + pow(v0[2] - v1[2],2));
+        float *v0 = &system->v[3*ip0];
+        float *v1 = &system->v[3*ip1];
+        float v_rel = sqrt(pow(v0[0] - v1[0],2) + pow(v0[1] - v1[1],2) + pow(v0[2] - v1[2],2));
 
         if( v_rel > crm ) {         // If relative speed larger than crm,
             crm = v_rel;            // then reset crm to larger value
@@ -151,8 +151,8 @@ double Cell::calculate_kinetic_energy() {
     return kinetic_energy;
 }
 
-vector<double>& Cell::update_average_velocity() {
-    vector<double> new_average_velocity(3,0);
+vector<float>& Cell::update_average_velocity() {
+    vector<float> new_average_velocity(3,0);
 
     for(int n=0; n<num_molecules; n++) {
         int index = molecules[n];
