+++ code/base_code/program/dsmc/cell.h
@@ -23,7 +23,7 @@ public:
     double collision_coefficient;
     double origin[3];
     int index_vector[3];
-    vector<double> average_velocity;
+    vector<float> average_velocity;
 
     System *system;
     int *molecules;
@@ -39,11 +39,11 @@ public:
     unsigned long prepare();
     void resize(int n);
     int collide(Random *rnd);
-    inline void collide_molecules(double *v0, double *v1, const double &v_rel, Random *rnd);
+    inline void collide_molecules(float *v0, float *v1, const float &v_rel, Random *rnd);
     void update_volume();
     void add_molecule(const int &molecule_index, unsigned long *index_in_cell, unsigned long *cell_index);
     void remove_molecule(const int &molecule_index, unsigned long *index_in_cell);
-    vector<double> &update_average_velocity();
+    vector<float> &update_average_velocity();
 
     double calculate_kinetic_energy();
     static bool cmp(Cell *c1, Cell *c2);
