+++ code/base_code/program/dsmc/colliderbase.h
@@ -8,5 +8,5 @@ protected:
 public:
     unsigned long num_collisions;
     ColliderBase();
-    virtual void collide(Random *rnd, double *v, float *normal_vector, float *tangent_vector_1, float *tangent_vector_2);
+    virtual void collide(Random *rnd, float *v, float *normal_vector, float *tangent_vector_1, float *tangent_vector_2);
 };
